// libs/shared/data-access/short-entity-list/src/lib/enums/short-entity-name.enum.ts
export enum ShortEntityNameEnum {
    Affiliates = 'affiliates',
    ActiveAffiliate = 'activeAffiliate',
    Advertisers = 'advertisers',
    Offers = 'offers',
    GeoNames = 'geoNames',
    LandingPage = 'landingdPage',
    Sponsor = 'sponsor',
    Managers = 'managers',
    PaymentMethods = 'paymentMethods',
    OperatingSystem = 'operatingSystem',
    Role = 'role',
    BaseRole = 'baseRole'
}